I made significant organizational changes.
All of the code is moved into a "sampex_microburst_widths"
entry point. The setup.py, config.cfg, and requirements.txt
files instruct pip to install this project in a virtual environent.

Install with python3 -m pip install -r requirements.txt file in the
virutal environment.

A second feature I added is I removed the dirs.py scrips and replaced
it with config.py file that lives in the nested sampex_microburst_widths
project folder. This script is writen automatically by calling this project
as a module which calls __main__.py. This can be set with
"python3 -m sampex_microburst_widths init".